[Refactor] ItemType enum 파일 위치 옮김 #203
