+++ src/main/java/com/yzgeneration/evc/domain/chat/implement/ChatRoomManager.java
@@ -4,7 +4,7 @@ import com.yzgeneration.evc.domain.chat.infrastructure.ChatMemberRepository;
 import com.yzgeneration.evc.domain.chat.infrastructure.ChatRoomRepository;
 import com.yzgeneration.evc.domain.chat.model.ChatMember;
 import com.yzgeneration.evc.domain.chat.model.ChatRoom;
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 import com.yzgeneration.evc.exception.CustomException;
 import com.yzgeneration.evc.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
