+++ src/main/java/com/yzgeneration/evc/domain/chat/infrastructure/ChatMessageRepository.java
@@ -4,7 +4,7 @@ import com.yzgeneration.evc.common.dto.SliceResponse;
 import com.yzgeneration.evc.domain.chat.dto.ChatMessageSliceResponse;
 import com.yzgeneration.evc.domain.chat.dto.ChatRoomListResponse;
 import com.yzgeneration.evc.domain.chat.model.ChatMessage;
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 
 import java.time.LocalDateTime;
