+++ src/main/java/com/yzgeneration/evc/domain/chat/infrastructure/ChatMessageRepositoryImpl.java
@@ -6,7 +6,7 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.yzgeneration.evc.common.dto.SliceResponse;
 import com.yzgeneration.evc.domain.chat.dto.*;
 import com.yzgeneration.evc.domain.chat.model.ChatMessage;
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 import com.yzgeneration.evc.domain.image.infrastructure.entity.QProfileImageEntity;
 
 import com.yzgeneration.evc.domain.member.infrastructure.QMemberEntity;
