+++ src/main/java/com/yzgeneration/evc/domain/chat/infrastructure/ChatRoomEntity.java
@@ -1,7 +1,7 @@
 package com.yzgeneration.evc.domain.chat.infrastructure;
 
 import com.yzgeneration.evc.domain.chat.model.ChatRoom;
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 import jakarta.persistence.*;
 import lombok.*;
