+++ src/main/java/com/yzgeneration/evc/domain/chat/infrastructure/ChatRoomRepository.java
@@ -1,7 +1,7 @@
 package com.yzgeneration.evc.domain.chat.infrastructure;
 
 import com.yzgeneration.evc.domain.chat.model.ChatRoom;
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 
 import java.util.Optional;
