+++ src/main/java/com/yzgeneration/evc/domain/delivery/controller/DeliveryController.java
@@ -4,7 +4,7 @@ import com.yzgeneration.evc.common.dto.SliceResponse;
 import com.yzgeneration.evc.domain.delivery.dto.DeliveryPrepare;
 import com.yzgeneration.evc.domain.delivery.model.DeliveryView;
 import com.yzgeneration.evc.domain.delivery.service.DeliveryService;
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 import com.yzgeneration.evc.external.delivery.GetKakaoMobilityOrder;
 import com.yzgeneration.evc.external.delivery.KakaoMobilityOrderResponse;
 import com.yzgeneration.evc.security.MemberPrincipal;
