+++ src/main/java/com/yzgeneration/evc/domain/delivery/model/Delivery.java
@@ -1,9 +1,8 @@
 package com.yzgeneration.evc.domain.delivery.model;
 
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.ToString;
 
 @Getter
 @Builder
