+++ src/main/java/com/yzgeneration/evc/domain/image/implement/ItemImageAppender.java
@@ -1,6 +1,6 @@
 package com.yzgeneration.evc.domain.image.implement;
 
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 import com.yzgeneration.evc.domain.image.model.ItemImage;
 import com.yzgeneration.evc.domain.image.service.port.ItemImageRepository;
 import lombok.RequiredArgsConstructor;
