+++ src/main/java/com/yzgeneration/evc/domain/image/infrastructure/entity/ItemImageEntity.java
@@ -1,6 +1,6 @@
 package com.yzgeneration.evc.domain.image.infrastructure.entity;
 
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 import com.yzgeneration.evc.domain.image.model.ItemImage;
 import jakarta.persistence.*;
 import lombok.*;
