+++ src/main/java/com/yzgeneration/evc/domain/image/model/ItemImage.java
@@ -1,6 +1,6 @@
 package com.yzgeneration.evc.domain.image.model;
 
-import com.yzgeneration.evc.domain.image.enums.ItemType;
+import com.yzgeneration.evc.domain.item.enums.ItemType;
 import lombok.Builder;
 import lombok.Getter;
